Integrate unit comparisons into main bundle
