+++ index.d.ts
@@ -8,7 +8,7 @@ declare namespace dayjs {
   export type OptionType = { locale: string }
 
   export type UnitType = 'millisecond' | 'second' | 'minute' | 'hour' | 'day' | 'week' | 'month' | 'quarter' | 'year' | 'date'
-  
+
   interface DayjsObject {
     years: number
     months: number
@@ -18,68 +18,68 @@ declare namespace dayjs {
     seconds: number
     milliseconds: number
   }
-  
+
   class Dayjs {
     constructor (config?: ConfigType)
-  
+
     clone(): Dayjs
-  
+
     isValid(): boolean
-  
+
     year(): number
-  
+
     month(): number
-  
+
     date(): number
-  
+
     day(): number
-  
+
     hour(): number
-  
+
     minute(): number
-  
+
     second(): number
-  
+
     millisecond(): number
-  
+
     set(unit: UnitType, value: number): Dayjs
-  
+
     add(value: number, unit: UnitType): Dayjs
-  
+
     subtract(value: number, unit: UnitType): Dayjs
-  
+
     startOf(unit: UnitType): Dayjs
-  
+
     endOf(unit: UnitType): Dayjs
-  
+
     format(template?: string): string
-  
+
     diff(dayjs: Dayjs, unit: UnitType, float?: boolean): number
-  
+
     valueOf(): number
-  
+
     unix(): number
-  
+
     daysInMonth(): number
-  
+
     toDate(): Date
-  
+
     toArray(): number[]
-  
+
     toJSON(): string
-  
+
     toISOString(): string
-  
+
     toObject(): DayjsObject
-  
+
     toString(): string
-  
-    isBefore(dayjs: Dayjs): boolean
-  
-    isSame(dayjs: Dayjs): boolean
-  
-    isAfter(dayjs: Dayjs): boolean
-  
+
+    isBefore(dayjs: Dayjs, unit?: UnitType): boolean
+
+    isSame(dayjs: Dayjs, unit?: UnitType): boolean
+
+    isAfter(dayjs: Dayjs, unit?: UnitType): boolean
+
     isLeapYear(): boolean
 
     locale(arg1: any, arg2?: any): Dayjs
