+++ src/plugin/isSameOrAfter/index.js
@@ -1,0 +1,5 @@
+export default (o, c) => {
+  c.prototype.isSameOrAfter = function (that, units) {
+    return this.isSame(that, units) || this.isAfter(that, units)
+  }
+}
