+++ src/plugin/isSameOrBefore/index.js
@@ -1,0 +1,5 @@
+export default (o, c) => {
+  c.prototype.isSameOrBefore = function (that, units) {
+    return this.isSame(that, units) || this.isBefore(that, units)
+  }
+}
